requirements.txt
matplotlib==3.5.1
Pillow==9.0.0
pre-commit==2.17.0